+++ protocols/kad/src/behaviour.rs
@@ -214,14 +214,28 @@ impl Default for KademliaConfig {
 }
 
 impl KademliaConfig {
+    /// Sets custom protocol names.
+    ///
+    /// Kademlia nodes only communicate with other nodes using the same protocol
+    /// name. Using custom name(s) therefore allows to segregate the DHT from
+    /// others, if that is desired.
+    ///
+    /// More than one protocol name can be supplied. In this case the node will
+    /// be able to talk to other nodes supporting any of the provided names.
+    /// Multiple names must be used with caution to avoid network partitioning.
+    pub fn set_protocol_names(&mut self, names: Vec<Cow<'static, [u8]>>) -> &mut Self {
+        self.protocol_config.set_protocol_names(names);
+        self
+    }
+
     /// Sets a custom protocol name.
     ///
     /// Kademlia nodes only communicate with other nodes using the same protocol
     /// name. Using a custom name therefore allows to segregate the DHT from
     /// others, if that is desired.
+    #[deprecated(since = "0.40.0", note = "use `set_protocol_names()` instead")]
     pub fn set_protocol_name(&mut self, name: impl Into<Cow<'static, [u8]>>) -> &mut Self {
-        self.protocol_config.set_protocol_name(name);
-        self
+        self.set_protocol_names(std::iter::once(name.into()).collect())
     }
 
     /// Sets the timeout for a single query.
@@ -403,8 +417,8 @@ where
     }
 
     /// Get the protocol name of this kademlia instance.
-    pub fn protocol_name(&self) -> &[u8] {
-        self.protocol_config.protocol_name()
+    pub fn protocol_names(&self) -> &[Cow<'static, [u8]>] {
+        self.protocol_config.protocol_names()
     }
 
     /// Creates a new `Kademlia` network behaviour with the given configuration.